refactor: move source assign package under split (#18253)
