+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/source/HoodieSource.java
@@ -25,8 +25,8 @@ import org.apache.hudi.common.model.HoodieTableType;
 import org.apache.hudi.common.table.HoodieTableMetaClient;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.ValidationUtils;
-import org.apache.hudi.source.assign.HoodieSplitAssigner;
-import org.apache.hudi.source.assign.HoodieSplitAssigners;
+import org.apache.hudi.source.split.assign.HoodieSplitAssigner;
+import org.apache.hudi.source.split.assign.HoodieSplitAssigners;
 import org.apache.hudi.configuration.FlinkOptions;
 import org.apache.hudi.exception.HoodieException;
 import org.apache.hudi.source.enumerator.HoodieContinuousSplitEnumerator;
