+++ hudi-flink-datasource/hudi-flink/src/test/java/org/apache/hudi/source/split/TestDefaultHoodieSplitProvider.java
@@ -20,7 +20,9 @@ package org.apache.hudi.source.split;
 
 import org.apache.hudi.common.util.Option;
 
-import org.apache.hudi.source.assign.HoodieSplitNumberAssigner;
+import org.apache.hudi.source.split.assign.HoodieSplitBucketAssigner;
+import org.apache.hudi.source.split.assign.HoodieSplitNumberAssigner;
+import org.apache.hudi.source.split.assign.DefaultHoodieSplitAssigner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -540,8 +542,8 @@ public class TestDefaultHoodieSplitProvider {
   @Test
   public void testGetNextWithDefaultAssigner() {
     // Test with DefaultHoodieSplitAssigner instead of NumberAssigner
-    org.apache.hudi.source.assign.DefaultHoodieSplitAssigner assigner =
-        new org.apache.hudi.source.assign.DefaultHoodieSplitAssigner(3);
+    DefaultHoodieSplitAssigner assigner =
+        new DefaultHoodieSplitAssigner(3);
     DefaultHoodieSplitProvider provider = new DefaultHoodieSplitProvider(assigner);
 
     List<HoodieSourceSplit> splits = new ArrayList<>();
@@ -561,8 +563,8 @@ public class TestDefaultHoodieSplitProvider {
   @Test
   public void testGetNextWithBucketAssigner() {
     // Test with HoodieSplitBucketAssigner
-    org.apache.hudi.source.assign.HoodieSplitBucketAssigner assigner =
-        new org.apache.hudi.source.assign.HoodieSplitBucketAssigner(4);
+    HoodieSplitBucketAssigner assigner =
+        new HoodieSplitBucketAssigner(4);
     DefaultHoodieSplitProvider provider = new DefaultHoodieSplitProvider(assigner);
 
     // Create splits with bucket-encoded file IDs
